Expense_Tracker/requirements.txt
Jinja2==3.1.2
fastapi==0.102.0
pydantic==1.10.10
PyMySQL==1.0.2
regex==2021.7.6
SQLAlchemy==2.0.21
uvicorn[standard]==0.17.*
typing==3.5
alembic==1.7.7
python-dotenv==0.10.1
cryptography==41.0.4
httpx==0.25.0
pytest==7.2.0
python-http-client==3.3.7
clients==1.4
requests==2.31.0